Align representations of board-state
The board-state is interacted with as a vector of vectors, where each
sub-vector contains the values of a particular column in the board. This
is mentioned briefly in a comment, and can be seen with the `move`
function, but the `cols` and `rows` functions didn't align with this
world view.

Instead of altering the code to interact with board-state in a
row-then-column order, I decided to swap the definitions of `cols` and
`rows`, and then add a docstring to the `empty-board` var calling out
this distinction. I certainly read it as row-then-column at first, and
calling attention to it in this way was the smallest code change
necessary.