Add initial take on generating all possible game states
This takes the approach of enumerating every possible combination of
moves (something like 184MM possibilities), and doesn't account for
early-exit win-states.

So it's incredibly brute-force, but lays a good foundation for what the
zippers are for and how/why we might use them in generating this tree.
I'm still curious about generating final board-states as opposed to move
lists, as that should cut down on the space and time complexity
significantly.