Add partial answer for Challenge 3
While working on generating all board-states (as opposed to permuting
all moves, which is what I did in Challenge 2), I realized that I'd
probably wandered over the "few hours of effort" suggested in the
README, so I've decided to simply record my ideas, my approach, and the
code I've written so far to generate the board states. Perhaps that is
sufficient for this task.